backend/lightvisiontest/requirements.txt
Django == 5.0.4
asgiref == 3.7.2
django-rest-framework == 0.1.0
djangorestframework == 3.14.0
psycopg2 == 2.9.9
setuptools == 69.0.3
tzdata == 2023.4
wheel == 0.42.0
pillow == 10.2.0
poster3 == 0.8.1
django-filter == 24.2
requests == 2.31.0